Add basic functional.
One time links, but files are still stored in FS without removing

+++ app/app.py
@@ -1,4 +1,7 @@
+import threading
 import urllib
+from time import sleep
+import garmage_collector
 
 import flask
 from flask import Flask, render_template, request, redirect, url_for, send_file
@@ -7,11 +10,13 @@ import os
 import uuid
 from datetime import datetime, timedelta
 from config import server_config
+from flask import after_this_request
+
 
 app = Flask(__name__)
 app.config['DEBUG'] = server_config['DEBUG']
 app.config['TEMP_LINK_DURATION'] = timedelta(hours=1)  # Время действия одноразовой ссылки
-
+app.config['UPLOAD_DIR'] = server_config['UPLOAD_DIR']
 
 
 def save_file(file, filename):
@@ -22,10 +27,18 @@ def save_file(file, filename):
 def upload_note():
     if request.method == 'POST':
         uploaded_text = request.form['text_note'].replace('\r','')
-        temp_uuid = memcached_toolkit.create_note(uploaded_text)
-        temp_link = url_for('download', temp_uuid=temp_uuid, _external=True)
+        note_uuid = memcached_toolkit.create_note(uploaded_text)
+        temp_link = url_for('download', temp_uuid=note_uuid, _external=True)
+        return render_template('show_temp_link.html', temp_link=temp_link)
 
-        # return f"Ваша одноразовая ссылка: <br /> {url_for('download', temp_uuid=temp_uuid, _external=True)}"
+
+
+@app.route('/file', methods=['POST'])
+def upload_file():
+    if request.method == 'POST':
+        uploaded_file = request.files['file']
+        file_uuid = memcached_toolkit.create_file(uploaded_file)
+        temp_link = url_for('download', temp_uuid=file_uuid, _external=True)
         return render_template('show_temp_link.html', temp_link=temp_link)
 
 @app.route('/', methods=['GET', 'POST'])
@@ -39,11 +52,21 @@ def download(temp_uuid):
     if file_meta is None:
         return render_template('404.html'), 404
 
-    memcached_toolkit.remove_entry(temp_uuid)
     if file_meta['is_note']:
-        note_data = open(os.path.join('uploads', temp_uuid), 'r', encoding='utf-8').read()
+        note_data = open(os.path.join('.\\uploads', temp_uuid), 'r', encoding='utf-8').read()
+        memcached_toolkit.remove_entry(temp_uuid)
         return render_template('note_display.html', note_data=note_data)
-    return 'Heh'
+
+    else:
+        if request.args.get('confirm') is None:
+            return render_template('file_display.html')
+        memcached_toolkit.remove_entry(temp_uuid)
+
+        print('before_response')
+
+        return send_file(os.path.join('.\\uploads', temp_uuid), download_name=file_meta['real_filename'], as_attachment=True)
+
 
 if __name__ == '__main__':
+    threading.Thread(target=garmage_collector.cleanup_files).start()
     app.run()
